Crate witx

Modules

parser
Witx syntax parsing from SExprs

phases
Paths to witx documents for various proposal phases

polyfill
Calculate required polyfill between interfaces

Structs

Case
One of a number of possible types that a Variant can take.

Constant

DocValidation

Document

Filesystem

HandleDatatype

Id

InterfaceFunc

InterfaceFuncParam

Location
Location in the source text

MockFs

Module

ModuleImport

NamedType

RecordDatatype
A struct-like value with named fields.

RecordMember

RecordMemberLayout

SizeAlign

Variant
A type which represents how values can be one of a set of possible cases.

Enums

Abi
Possible ABIs for interface functions to have.

BuiltinType

Definition

Entry

Instruction

IntRepr

ModuleDefinition

ModuleEntry

ModuleImportVariant

RecordKind
Different kinds of records used for hinting various language-specific types.

RepEquality

SExpr

Type
Structure of all possible interface types.

TypeRef

ValidationError

WasmType
Enumerates wasm types used by interface types when lowering/lifting.

WitxError

Traits

Bindgen
Trait for language implementors to use to generate glue code between native WebAssembly signatures and interface types signatures.

Documentation
Enables generating Markdown formatted content.

Layout

Representable

WitxIo

Functions

load
Load a witx document from the filesystem

parse
Parse a witx document from a str. (use ...) directives are not permitted.