Crate witx

Modules

phases
Paths to witx documents for various proposal phases

polyfill
Calculate required polyfill between interfaces

Structs

CoreFuncType
Describes the WebAssembly signature of a function

CoreParamType
A parameter in the WebAssembly type of a function.

Document

EnumDatatype

EnumVariant

Filesystem

FlagsDatatype

FlagsMember

HandleDatatype

Id

IntConst

IntDatatype

InterfaceFunc

InterfaceFuncParam

Location
Location in the source text

MockFs

Module

ModuleImport

NamedType

SizeAlign

StructDatatype

StructMember

StructMemberLayout

UnionDatatype

UnionLayout

UnionVariant

Enums

AtomType
Enumerates the types permitted for function arguments in the WebAssembly spec

BuiltinType

CoreParamSignifies
Enumerates the sort of relationship an WebAssembly parameter to an interface function parameter.

DeclSyntax

Definition

Entry

IntRepr

InterfaceFuncParamPosition

ModuleDefinition

ModuleEntry

ModuleImportVariant

RepEquality

SExpr

Type

TypePassedBy
Enumerates the strategies which may be used to pass a datatype as an argument

TypeRef

ValidationError

WitxError

Traits

Documentation
Enables generating Markdown formatted content.

Layout

Representable

WitxIo

Functions

load
Load a witx document from the filesystem

parse
Parse a witx document from a str. (use ...) directives are not permitted.